+++ graph/node_test.go
@@ -903,6 +903,35 @@ func TestGather(t *testing.T) {
 
 }
 
+func TestGatherSlices(t *testing.T) {
+	testFuncOneInput(t, "GatherSlices(input, slicedAxes={1}, start={{0}, {1}, {0}}, sizes={1})",
+		func(g *Graph) (input, output *Node) {
+			input = IotaFull(g, shapes.Make(shapes.F32, 4, 5))
+			start := Const(g, [][]int32{{0}, {1}, {0}}) // Slice from rows 0, 2 and 0 of each example in the batch.
+			sizes := []int{1}                           // Take only one row per start.
+			output = GatherSlices(input, []int{0}, start, sizes)
+			return
+		}, [][][]float32{{{0, 1, 2, 3, 4}}, {{5, 6, 7, 8, 9}}, {{0, 1, 2, 3, 4}}})
+
+	testFuncOneInput(t, "GatherSlices(input, slicedAxes={0}, start={{0}, {1}}, sizes={2})",
+		func(g *Graph) (input, output *Node) {
+			input = IotaFull(g, shapes.Make(shapes.F32, 4, 3))
+			start := Const(g, [][]int32{{0}, {1}}) // Slice from rows 0 and 1.
+			sizes := []int{2}                      // Take two rows per start.
+			output = GatherSlices(input, []int{0}, start, sizes)
+			return
+		}, [][][]float32{{{0, 1, 2}, {3, 4, 5}}, {{3, 4, 5}, {6, 7, 8}}})
+
+	testFuncOneInput(t, "GatherSlices(input, slicedAxes={0,1}, start={1, 1}, sizes={2, 3})",
+		func(g *Graph) (input, output *Node) {
+			input = IotaFull(g, shapes.Make(shapes.F32, 4, 10))
+			start := Const(g, []int32{1, 1}) // Slice in middle of matrix.
+			sizes := []int{2, 3}             // Take a sub-matrix
+			output = GatherSlices(input, []int{0, 1}, start, sizes)
+			return
+		}, [][]float32{{11, 12, 13}, {21, 22, 23}})
+}
+
 func TestIndicesForShape(t *testing.T) {
 	manager := BuildManager().MustDone()
 	g := manager.NewGraph("")
@@ -1122,106 +1151,6 @@ func TestBatchNormInferenceXLA(t *testing.T) {
 			{14.5, 109, 1003.79474},
 			{15.5, 111, 1004.7434},
 		})
-}
-
-func TestEinsum(t *testing.T) {
-	graphtest.RunTestGraphFn(t, "EinsumMatrixMul",
-		func(g *Graph) (inputs, outputs []*Node) {
-			lhs := IotaFull(g, MakeShape(F32, 2, 4))
-			lhs = OnePlus(lhs)
-			rhs := MulScalar(Ones(g, MakeShape(F32, 4, 3)), 0.1)
-			inputs = []*Node{lhs, rhs}
-			outputs = []*Node{Einsum("ij,jk->ik", lhs, rhs)}
-			return
-		}, []any{[][]float32{{1, 1, 1}, {2.6, 2.6, 2.6}}}, slices.Epsilon)
-	graphtest.RunTestGraphFn(t, "EinsumDotProduct",
-		func(g *Graph) (inputs, outputs []*Node) {
-			lhs := IotaFull(g, MakeShape(F32, 4))
-			lhs = OnePlus(lhs)
-			rhs := MulScalar(Ones(g, MakeShape(F32, 4)), 0.1)
-			inputs = []*Node{lhs, rhs}
-			outputs = []*Node{Einsum("i,i->", lhs, rhs)}
-			return
-		}, []any{float32(1)}, slices.Epsilon)
-	graphtest.RunTestGraphFn(t, "EinsumOuterProduct",
-		func(g *Graph) (inputs, outputs []*Node) {
-			lhs := OnePlus(IotaFull(g, MakeShape(F32, 4)))
-			rhs := OnePlus(IotaFull(g, MakeShape(F32, 3)))
-			inputs = []*Node{lhs, rhs}
-			outputs = []*Node{Einsum("i,j->ij", lhs, rhs)}
-			return
-		}, []any{[][]float32{{1, 2, 3}, {2, 4, 6}, {3, 6, 9}, {4, 8, 12}}}, slices.Epsilon)
-	graphtest.RunTestGraphFn(t, "EinsumBatchMatrixMul",
-		func(g *Graph) (inputs, outputs []*Node) {
-			lhs := IotaFull(g, MakeShape(F32, 5, 2, 4))
-			lhs = OnePlus(lhs)
-			rhs := MulScalar(Ones(g, MakeShape(F32, 5, 4, 3)), 0.1)
-			inputs = []*Node{lhs, rhs}
-			outputs = []*Node{Einsum("bij,bjk->bik", lhs, rhs)}
-			return
-		}, []any{[][][]float32{
-			{{1, 1, 1}, {2.6, 2.6, 2.6}},
-			{{4.2, 4.2, 4.2}, {5.8, 5.8, 5.8}},
-			{{7.4, 7.4, 7.4}, {9, 9, 9}},
-			{{10.6, 10.6, 10.6}, {12.2, 12.2, 12.2}},
-			{{13.8, 13.8, 13.8}, {15.4, 15.4, 15.4}}},
-		}, slices.Epsilon)
-	graphtest.RunTestGraphFn(t, "EinsumBatchMatrixMulTransposed",
-		func(g *Graph) (inputs, outputs []*Node) {
-			lhs := IotaFull(g, MakeShape(F32, 5, 2, 4))
-			lhs = OnePlus(lhs)
-			rhs := MulScalar(Ones(g, MakeShape(F32, 5, 4, 3)), 0.1)
-			inputs = []*Node{lhs, rhs}
-			outputs = []*Node{Einsum("bij,bjk->ikb", lhs, rhs)}
-			return
-		}, []any{[][][]float32{
-			{{1, 4.2, 7.4, 10.6, 13.8},
-				{1, 4.2, 7.4, 10.6, 13.8},
-				{1, 4.2, 7.4, 10.6, 13.8}},
-			{{2.6, 5.8, 9, 12.2, 15.4},
-				{2.6, 5.8, 9, 12.2, 15.4},
-				{2.6, 5.8, 9, 12.2, 15.4}},
-		}}, slices.Epsilon)
-}
-
-func TestLowerTriangular(t *testing.T) {
-	graphtest.RunTestGraphFn(t, "LowerTriangular(3)",
-		func(g *Graph) (inputs, outputs []*Node) {
-			inputs = []*Node{}
-			outputs = []*Node{LowerTriangular(g, 3)}
-			return
-		}, []any{[][]bool{
-			{true, false, false},
-			{true, true, false},
-			{true, true, true},
-		}}, -1)
-}
-
-func TestUpperTriangular(t *testing.T) {
-	graphtest.RunTestGraphFn(t, "LowerTriangular(3)",
-		func(g *Graph) (inputs, outputs []*Node) {
-			inputs = []*Node{}
-			outputs = []*Node{UpperTriangular(g, 3)}
-			return
-		}, []any{[][]bool{
-			{true, true, true},
-			{false, true, true},
-			{false, false, true},
-		}}, -1)
-}
-
-func TestDiagonalWithValue(t *testing.T) {
-	graphtest.RunTestGraphFn(t, "DiagonalWithValue(5, 3)",
-		func(g *Graph) (inputs, outputs []*Node) {
-			value := Const(g, 5.0)
-			inputs = []*Node{value}
-			outputs = []*Node{DiagonalWithValue(value, 3)}
-			return
-		}, []any{[][]float64{
-			{5, 0, 0},
-			{0, 5, 0},
-			{0, 0, 5},
-		}}, -1)
 }
 
 func TestSqueeze(t *testing.T) {